T4ARegistrator

Documentation for T4ARegistrator.

T4ARegistrator.PkgBump.bump

T4ARegistrator.PkgBump.bumpmajor

T4ARegistrator.PkgBump.bumpminor

T4ARegistrator.PkgBump.bumppatch

T4ARegistrator.PkgBump.updatemajor

T4ARegistrator.PkgBump.updatemajor!

T4ARegistrator.PkgBump.updateminor

T4ARegistrator.PkgBump.updateminor!

T4ARegistrator.PkgBump.updatepatch

T4ARegistrator.PkgBump.updatepatch!

T4ARegistrator.PkgBump.updateversion

T4ARegistrator.PkgBump.updateversion!

T4ARegistrator.register

T4ARegistrator.register — Function

register(targetpkg::Module)
register()

Register package::Module. If package is omitted, register the package in the currently active project or in the current directory in case the active project is not a package.

T4ARegistrator.PkgBump.bump — Method

bump(mode::Symbol; commit=true, push=true)

Bumps the version of the current active project according to mode, commits the change to a new branch, and pushes the branch to the remote repository.

T4ARegistrator.PkgBump.bumpmajor — Method

bumpmajor(;kwargs)

Bump the major version of the current active project, commit, and push the changes.

T4ARegistrator.PkgBump.bumpminor — Method

bumpminor(;kwargs)

Bump the minor version of the current active project, commit, and push the changes.

T4ARegistrator.PkgBump.bumppatch — Method

bumppatch(;kwargs)

Bump the patch version of the current active project, commit, and push the changes.

T4ARegistrator.PkgBump.updatemajor! — Method

updatemajor!(project::Pkg.Types.Project, project_file)

Increment the major version of the given project and write the changes to the project_file.

T4ARegistrator.PkgBump.updatemajor — Method

updatemajor(project_file::AbstractString)

Update the major version of the project defined in project_file.

T4ARegistrator.PkgBump.updateminor! — Method

updateminor!(project::Pkg.Types.Project, project_file::AbstractString)

Increment the minor version of the given project and write the changes to the project_file.

T4ARegistrator.PkgBump.updateminor — Method

updateminor(project_file::AbstractString)

Update the minor version of the project defined in project_file.

T4ARegistrator.PkgBump.updatepatch! — Method

updatepatch!(project::Pkg.Types.Project, project_file::AbstractString)

Increment the patch version of the given project and write the changes to the project_file.

T4ARegistrator.PkgBump.updatepatch — Method

updatepatch(project_file::AbstractString)

Update the patch version of the project defined in project_file.

T4ARegistrator.PkgBump.updateversion! — Method

updateversion!(project::Pkg.Types.Project, project_file::AbstractString, mode::Symbol)

Update the version of the given project file according to the specified mode (:patch, :minor, :major). The new version is written directly to the project_file.

T4ARegistrator.PkgBump.updateversion — Method

updateversion(project_file::AbstractString, mode::Symbol) -> Pkg.Types.Project

Read the project from project_file, update its version according to mode, and write the changes back. Returns the updated project.